# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Early Dementia Detection via Speech Analysis API
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * Accepts audio file, transcribes it, and performs cognitive analysis
 * @summary Transcribe and analyze audio
 */
export const transcribeAudioBodyLanguageDefault = `en`;

export const TranscribeAudioBody = zod.object({
  audio: zod.instanceof(File),
  language: zod.enum(["en", "te"]).default(transcribeAudioBodyLanguageDefault),
  userId: zod.string().nullish(),
});

export const TranscribeAudioResponse = zod.object({
  transcript: zod.string(),
  riskLevel: zod.enum(["Normal", "Mild Cognitive Impairment", "High Risk"]),
  riskScore: zod.number(),
  confidence: zod.number(),
  features: zod.object({
    wordCount: zod.number(),
    uniqueWordCount: zod.number(),
    vocabularyDiversity: zod.number(),
    avgSentenceLength: zod.number(),
    pauseCount: zod.number(),
    repetitionCount: zod.number(),
    repetitionRate: zod.number(),
    fillerWordCount: zod.number(),
    coherenceScore: zod.number(),
  }),
  problematicPatterns: zod.array(zod.string()),
  recommendations: zod.array(zod.string()),
  language: zod.string(),
});

/**
 * Analyze transcribed text for dementia risk markers
 * @summary Analyze text for cognitive markers
 */
export const analyzeTextBodyLanguageDefault = `en`;

export const AnalyzeTextBody = zod.object({
  text: zod.string(),
  language: zod.enum(["en", "te"]).default(analyzeTextBodyLanguageDefault),
  userId: zod.string().nullish(),
});

export const AnalyzeTextResponse = zod.object({
  transcript: zod.string(),
  riskLevel: zod.enum(["Normal", "Mild Cognitive Impairment", "High Risk"]),
  riskScore: zod.number(),
  confidence: zod.number(),
  features: zod.object({
    wordCount: zod.number(),
    uniqueWordCount: zod.number(),
    vocabularyDiversity: zod.number(),
    avgSentenceLength: zod.number(),
    pauseCount: zod.number(),
    repetitionCount: zod.number(),
    repetitionRate: zod.number(),
    fillerWordCount: zod.number(),
    coherenceScore: zod.number(),
  }),
  problematicPatterns: zod.array(zod.string()),
  recommendations: zod.array(zod.string()),
  language: zod.string(),
});

/**
 * @summary Get all reports for a user
 */
export const GetReportsQueryParams = zod.object({
  userId: zod.coerce.string().optional(),
});

export const GetReportsResponseItem = zod.object({
  id: zod.number(),
  userId: zod.string(),
  transcript: zod.string(),
  riskLevel: zod.string(),
  riskScore: zod.number(),
  confidence: zod.number(),
  features: zod.object({
    wordCount: zod.number(),
    uniqueWordCount: zod.number(),
    vocabularyDiversity: zod.number(),
    avgSentenceLength: zod.number(),
    pauseCount: zod.number(),
    repetitionCount: zod.number(),
    repetitionRate: zod.number(),
    fillerWordCount: zod.number(),
    coherenceScore: zod.number(),
  }),
  problematicPatterns: zod.array(zod.string()),
  recommendations: zod.array(zod.string()),
  language: zod.string(),
  createdAt: zod.date(),
});
export const GetReportsResponse = zod.array(GetReportsResponseItem);

/**
 * @summary Save a report
 */
export const CreateReportBody = zod.object({
  userId: zod.string(),
  transcript: zod.string(),
  riskLevel: zod.string(),
  riskScore: zod.number(),
  confidence: zod.number(),
  features: zod.object({
    wordCount: zod.number(),
    uniqueWordCount: zod.number(),
    vocabularyDiversity: zod.number(),
    avgSentenceLength: zod.number(),
    pauseCount: zod.number(),
    repetitionCount: zod.number(),
    repetitionRate: zod.number(),
    fillerWordCount: zod.number(),
    coherenceScore: zod.number(),
  }),
  problematicPatterns: zod.array(zod.string()),
  recommendations: zod.array(zod.string()),
  language: zod.string(),
});

/**
 * @summary Get a specific report
 */
export const GetReportParams = zod.object({
  id: zod.coerce.number(),
});

export const GetReportResponse = zod.object({
  id: zod.number(),
  userId: zod.string(),
  transcript: zod.string(),
  riskLevel: zod.string(),
  riskScore: zod.number(),
  confidence: zod.number(),
  features: zod.object({
    wordCount: zod.number(),
    uniqueWordCount: zod.number(),
    vocabularyDiversity: zod.number(),
    avgSentenceLength: zod.number(),
    pauseCount: zod.number(),
    repetitionCount: zod.number(),
    repetitionRate: zod.number(),
    fillerWordCount: zod.number(),
    coherenceScore: zod.number(),
  }),
  problematicPatterns: zod.array(zod.string()),
  recommendations: zod.array(zod.string()),
  language: zod.string(),
  createdAt: zod.date(),
});

/**
 * @summary Delete a report
 */
export const DeleteReportParams = zod.object({
  id: zod.coerce.number(),
});

/**
 * @summary Send message to cognitive health chatbot
 */
export const sendChatbotMessageBodyLanguageDefault = `en`;

export const SendChatbotMessageBody = zod.object({
  message: zod.string(),
  history: zod
    .array(
      zod.object({
        role: zod.enum(["user", "assistant"]),
        content: zod.string(),
      }),
    )
    .optional(),
  language: zod
    .enum(["en", "te"])
    .default(sendChatbotMessageBodyLanguageDefault),
});

export const SendChatbotMessageResponse = zod.object({
  reply: zod.string(),
  tips: zod.array(zod.string()),
});