# lib/api-zod/src/generated/types/analyzeTextBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Early Dementia Detection via Speech Analysis API
 * OpenAPI spec version: 0.1.0
 */
import type { AnalyzeTextBodyLanguage } from "./analyzeTextBodyLanguage";

export interface AnalyzeTextBody {
  text: string;
  language?: AnalyzeTextBodyLanguage;
  /** @nullable */
  userId?: string | null;
}