# lib/api-zod/src/generated/types/report.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Early Dementia Detection via Speech Analysis API
 * OpenAPI spec version: 0.1.0
 */
import type { SpeechFeatures } from "./speechFeatures";

export interface Report {
  id: number;
  userId: string;
  transcript: string;
  riskLevel: string;
  riskScore: number;
  confidence: number;
  features: SpeechFeatures;
  problematicPatterns: string[];
  recommendations: string[];
  language: string;
  createdAt: Date;
}